Inversion of control refactoring
